Refactor SentryService check in InitializerMiddleware
Replaced SentryService::isEnabled() method with SentryService::inititalize() inside InitializerMiddleware's process method. The reason for this change is to ensure that SentryService is not just enabled, but also properly initialized before the extension configuration is fetched and processed. This can prevent potential issues from arising if SentryService is enabled but not correctly initialized.